JsonSchema/requirements.txt
anyio==4.3.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==3.3.2
Choco==1.0.5
click==8.1.7
colorama==0.4.6
command-not-found==0.3
constantly==23.10.4
cors==1.0.1
cryptography==42.0.5
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.2.0
EZGmail==2022.10.10
fasteners==0.14.1
filelock==3.13.1
Flask==3.0.1
future==1.0.0
gevent==24.2.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
httplib2==0.20.2
httptools==0.6.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.6.4
incremental==22.10.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llfuse==1.3.8
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.4
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.8
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==2.2.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==4.25.2
psycopg==3.1.17
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
redis==5.0.3
referencing==0.32.1
reportlab==3.6.8
requests==2.31.0
requests-file==2.0.0
rpds-py==0.17.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
systemd-python==234
texttable==1.6.4
tldextract==5.1.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
wadllib==1.3.6
watchfiles==0.21.0
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.6.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.2
